Dockerfile: Don't copy .git to the container
The .git directory shouldn't be copied to the container.  It causes
unneeded rebuilds, because .git directory is constantly changing and the
git version history should not be needed inside the container.

Also update the submodules, because the packages cannot be installed
without the recent pyproject.toml changes in the repos.