Process configs with environment variables
Add environment variables for (most of) the configuration options in the
TOML files and modify the files so that they use the variables instead
of hardcoded values.  Add a script for processing those configuration
template files with envsubst so that the values are filled using the
current environment variables on container start up.

Also update the passari-workflow submodule to add support for
configuring the PostgreSQL connection with an URL.

Remove [logging] section from passari-workflow.toml, because it's not
actually used. (Only the [logging] section from passari.toml is used).