Set passlib log level to INFO for passari-web-ui
If general log level is set to DEBUG (as it is by default with the
"level=10" line in [logging] section of passari.toml), the flask app
outputs some surplus messages from passlib on each invocation.  Improve
this situation by setting the log level of passlib to INFO by default.

Also update the passari-web-ui submodule to bring in the feature which
allows settings log levels via CUSTOM_LOG_LEVELS option.