Dockerfile: Install siptools via wheels
Compile wheels for siptools packages and install them.

This is needed, because the siptools installed via RPMs is packaged for
Python 3.9, but we need it for Python 3.12.  The RPMs are still
installed to make sure that we have all the non-Python packages which
are needed by siptools (e.g. ffmpeg, VeraPDF, etc.).