Ignore keys directory
To specify key files for METS signing or SFTP connection they have to be
placed on some path that is available on the container.  This can now be
done by creating a directory named keys/ and placing them there.  That
directory will be available on the container as /app/keys, since
docker-compose mounts the workdir to /app as a volume.