Data directory for objects and archive
Create a data directory for the Passari object storage and archive and
make it a volume in the docker-compose setup.

Set the correct PASSARI_WORKFLOW_PACKAGE_DIR and
PASSARI_WORKFLOW_ARCHIVE_DIR variables so that passari-workflow is
configured with the correct paths.

Additionally create symbolic links MuseumObjects and MuseumObjectArchive
to the home directory, because passari's CLI tools use those as default
values.