ansible: Add requirements files
Add requirements.in and requirements.txt for listing/pinning the
packages needed to run Ansible.

It's also possible to use the Ansible packaged with various Linux
distros, but having a pinned version defined in requirements.txt helps
to find a version which is certainly compatible with the playbook files
here.