compile-requirements: Add hashes feature and improvements
Make it possible to generate requirements files with the hashes
included (for more secure installation), tune the pip-compile arguments
so that find-links argument like "-f ./wheels/" is not emitted to the
file and even unsafe packages like "setuptools" is version pinned.

Regenerate the requirements-dev.txt and requirements-siptools.txt files
with this change.