Fix requirements to be compilable again
After the dpres-mets-builder and dpres-siptools-ng were added as
dependencies to passari pip-compile didn't know where to find them and
therefore wasn't able to recompile the requirements*.txt files.

Fix this issue by adding a "-c siptools-deps.txt" line to the
requirements.in where the siptools-deps.txt contains the all of the
siptools dependencies and their download locations.

Also modify the requirements-siptools.in to use that siptools-deps.txt
so that the same deps don't have to be specified in two places.

Additionally, regenerate the requirements*.txt with these changes.