Update m2crypto to 0.46.2
It seems that the version 0.44.0 cannot be compiled on AlmaLinux 9
anymore.  Therefore update the m2crypto in our requirements.txt to
version 0.46.2.  Also pull the same change for the ansible submodule.

I'm not sure about the exact reason why 0.44.0 doesn't compile, but the
error message is:

    error: ‘fd_set’ has no member named ‘__fds_bits’

Same error was reported for the python-m2crypto Arch package [1].  There
is more info on the URLs found via a comment on that Arch package page:
[2] and [3].

[1] https://aur.archlinux.org/packages/python-m2crypto

[2] https://git.openembedded.org/meta-openembedded/commit/?id=f9158ce32fffa6f18eed4008c3295146c81d55ea

[3] https://github.com/swiftlang/swift/issues/69311